Explicitly capture analyzer-executor's dependency on grapl-common
Removing `grapl-common` from `analyzerlib`'s dependency list (because
it wasn't used) ended up breaking the container build of
`analyzer-executor`, because it was indirectly relying on that
dependency to fullfil its direct dependency on `grapl-common`.

Because our current Python organization and build system can't cleanly
deal with inner-repo dependencies like this, this commit modifies the
container build for `analyzer-executor` to compensate.

By moving its Dockerfile up a level to
`src/python/analyzer_executor.Dockerfile`, we can take advantage of
the fact that all the Python code is now available in the container
build context. We can then simply pull in the needed `grapl-common`
code and do a `pip install` directly in the Dockerfile.

This isn't ideal, since it doesn't capture the dependency in an
obvious or Python-standard manner. It does, however, present a minimal
solution that can be extended to other services as similar
dependency issues are uncovered. This in turn will allow us to
incrementally improve and tighten up our Python code structure. The
long-term goal is to not require such container build-time
shenanigans.

(For completeness, I initially tried moving the
`grapl-python-build/Dockerfile` up to `src/python` and adding _all_
the Python code to the container. This would have been nice because it
would immediately provide a similar method to resolve any other
inner-repo dependency problems. However, this would basically
guarantee build cache invalidation every time any Python code was
changed, since this container is at the root of every Python container
build chain. That'd be bad.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>